# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Dazeku API",
        "description": "API per la gestione della webapp Dazeku",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "dazeku.onrender.com",
    "basePath": "/",
...
    "schemes": ["https", "http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
...
        "/api/opera/getmany": {
            "get": {
                "tags": [
                    "opera"
                ],
                "parameters": [{
                    "in": "query",
                    "name": "titolo",
                    "schema": {
                        "type": "string"
                    }
                }],
                "summary": "Restituisci tutte le opere con visibilità pubblica che contengono nel titolo la stringa passata come parametro",
                "responses": {
                    "200": {
                        "description": "Le opere sono state trovate e restituite con successo",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Opera"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parametro titolo non specificato",
                        "schema": {
                            "type": "object",
                            "properties": {
                              "success": {
                                "type": "boolean"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": ["success", "message"]
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "query error"
                            }
                        }
                    },
                    "500": {
                        "description": "Il server non è riuscito a recuperare le opere",
                        "schema": {
                            "type": "object",
                            "properties": {
                              "success": {
                                "type": "boolean"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": ["success", "message"]
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "è avvenuto un errore"
                            }
                        }
                    }
                }
            }
        },
        "/api/opera/getmostsuccessful": {
            "get": {
                "tags": [
                    "opera"
                ],
                "summary": "Restituisce le 20 opere con più seguaci",
                "responses": {
                    "200": {
                        "description": "Le opere sono state trovate e restituite con successo",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Opera"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Il server non è riuscito a recuperare le opere",
                        "schema": {
                            "type": "object",
                            "properties": {
                              "success": {
                                "type": "boolean"
                              },
                              "message": {
                                "type": "string"
                              }
                            },
                            "required": ["success", "message"]
                        },
                        "examples": {
                            "application/json": {
                                "success": false,
                                "message": "è avvenuto un errore"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Opera": {
            "required": [
                "_id",
                "titolo",
                "generi",
                "autore",
                "autoreId",
                "sinossi",
                "nSeguaci",
                "listaCap",
                "visibile"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "titolo": {
                    "type": "string"
                },
                "generi": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "autore": {
                    "type": "string"
                },
                "autoreId": {
                    "type": "string"
                },
                "sinossi": {
                    "type": "string"
                },
                "nSeguaci": {
                    "type": "number"
                },
                "listaCap": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Capitolo"                        
                    }
                },
                "visibile": {
                    "type": "boolean"
                }
            }
        },
        "Capitolo": {
            "required": [
                "titolo",
                "operaId",
                "numero",
                "testo",
                "visibile"
            ],
            "properties": {
                "titolo": {
                    "type": "string"
                },
                "operaId": {
                    "type": "string"
                },
                "numero": {
                    "type": "number"
                },
                "testo": {
                    "type": "string"
                },
                "visibile": {
                    "type": "boolean"
                }
            }
        }
    }
}